requirements.txt
aiofiles==0.4.0
certifi==2019.11.28
chardet==3.0.4
h11==0.8.1
h2==3.1.1
hpack==3.0.0
hstspreload==2020.1.22
httptools==0.0.13
httpx==0.9.3
hyperframe==5.2.0
idna==2.8
multidict==4.7.4
python-chess==0.30.1
rfc3986==1.3.2
sanic==19.12.2
sniffio==1.1.0
ujson==1.35
uvloop==0.14.0
websockets==8.1
